ssz: move everything project specfic out of types module to ssz
