+++ ssz/build.gradle
@@ -1,6 +1,7 @@
 apply plugin: 'java-library'
 
 dependencies {
+    implementation project(':types')
     implementation 'net.consensys.cava:cava-ssz'
     implementation 'net.consensys.cava:cava-units'
