+++ types/build.gradle
@@ -14,7 +14,6 @@
 apply plugin: 'java-library'
 
 dependencies {
-  implementation project(':ssz')
   implementation 'com.google.guava:guava'
   implementation 'io.vertx:vertx-core'
   implementation 'net.consensys.cava:cava-ssz'
